cap nhat code

+++ UniCharity_BE/src/main/java/com/UniCharity/UniCharity/dto/request/CampaignCreateRequest.java
@@ -14,8 +14,8 @@ import java.time.Instant;
 public class CampaignCreateRequest {
     String title;
     String description;
-    BigDecimal targetAmount;
-    BigDecimal currentAmount;
+    long targetAmount;
+    long currentAmount;
     Instant startDate;
     Instant endDate;
     int departmentId;
